# Improve performance of switching initialized documents