--- a/src/js/jsx/sections/style/AppearancePanel.jsx
+++ b/src/js/jsx/sections/style/AppearancePanel.jsx
@@ -82,23 +82,23 @@ define(function (require, exports, module) {
             event.stopPropagation();
         },
 
-        shouldComponentUpdate: function (nextProps) {
-            if (this.props.disabled !== nextProps.disabled ||
-                this.props.active !== nextProps.active ||
-                this.props.hidden !== nextProps.hidden) {
-                return true;
-            }
-
-            if (!Immutable.is(this.props.document.layers.selected,
-                nextProps.document.layers.selected)) {
-                return true;
-            }
-
-            if (!nextProps.visible && !this.props.visible) {
+        shouldComponentUpdate: function (nextProps, nextState) {
+            // Only compare the existence of clipboard.
+            var clipboardChanged = !!this.state.clipboard !== !!nextState.clipboard,
+                selectedLayerChanged = !Immutable.is(this.props.document.layers.selected,
+                    nextProps.document.layers.selected);
+            
+            // If the panel is remaining invisible and the clipboard is unchanged, no need to re-render.
+            // - Copying layer effects will update buttons in the panel header.
+            // - Deselect layers will remove the panel
+            if (!nextProps.visible && !this.props.visible && !clipboardChanged && !selectedLayerChanged) {
                 return false;
             }
 
-            return true;
+            return this.props.disabled !== nextProps.disabled ||
+                this.props.visible !== nextProps.visible ||
+                clipboardChanged ||
+                selectedLayerChanged;
         },
 
         getStateFromFlux: function () {
@@ -177,8 +177,6 @@ define(function (require, exports, module) {
             var sectionClasses = classnames({
                 "section": true,
                 "appearance": true,
-                "section__active": this.props.active,
-                "section__sibling-collapsed": !this.props.visibleSibling,
                 "section__collapsed": !this.props.visible,
                 "appearance__mixed": !uniformLayerKind && hasSomeTextLayers && hasSomeVectorLayers
             });