--- a/src/style/sections/libraries/libraries-section.less
+++ b/src/style/sections/libraries/libraries-section.less
@@ -30,6 +30,7 @@
     position: relative;
     flex-grow: 1;
     flex-shrink: 0;
+    width: 100%;
 
     .section-container {
         padding: 0;