… src/style/sections/style/style-section.less
… src/style/shared/icons.less