@iwehrman 2016-02-10
This is excellent. I only have some very small suggestions. We should get this merged by the end of the day. Back to @shaoshing!